extract global variable generation from llvm-generator into dedicated module
